+++ cadc-storage-adapter-fs/src/intTest/java/org/opencadc/inventory/storage/fs/OpaquePutTxnTest.java
@@ -3,7 +3,7 @@
 *******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 **************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 *
-*  (c) 2020.                            (c) 2020.
+*  (c) 2021.                            (c) 2021.
 *  Government of Canada                 Gouvernement du Canada
 *  National Research Council            Conseil national de recherches
 *  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -79,16 +79,14 @@ import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.junit.Assert;
 import org.junit.Before;
-import org.opencadc.inventory.storage.test.StorageAdapterByteRangeTest;
+import org.opencadc.inventory.storage.test.StorageAdapterPutTxnTest;
 
 /**
- * Integration tests that interact with the file system. These tests require a file system
- * that supports posix extended attributes.
- * 
+ *
  * @author pdowler
  */
-public class OpaqueByteRangeTest extends StorageAdapterByteRangeTest {
-    private static final Logger log = Logger.getLogger(OpaqueByteRangeTest.class);
+public class OpaquePutTxnTest extends StorageAdapterPutTxnTest {
+    private static final Logger log = Logger.getLogger(OpaquePutTxnTest.class);
 
     static final int BUCKET_LEN = 2;
     static final File ROOT_DIR;
@@ -100,8 +98,8 @@ public class OpaqueByteRangeTest extends StorageAdapterByteRangeTest {
     }
     
     final OpaqueFileSystemStorageAdapter ofsAdapter;
-    
-    public OpaqueByteRangeTest() { 
+            
+    public OpaquePutTxnTest() { 
         super(new OpaqueFileSystemStorageAdapter(ROOT_DIR, BUCKET_LEN));
         this.ofsAdapter = (OpaqueFileSystemStorageAdapter) super.adapter;
 
@@ -113,7 +111,7 @@ public class OpaqueByteRangeTest extends StorageAdapterByteRangeTest {
     
     @Before
     public void cleanupBefore() throws IOException {
-        log.info("cleanupBefore: delete all content from " + ofsAdapter.contentPath);
+        log.info("cleanupBefore: " + ofsAdapter.contentPath.getParent());
         if (Files.exists(ofsAdapter.contentPath)) {
             Files.walkFileTree(ofsAdapter.contentPath, new SimpleFileVisitor<Path>() {
                 @Override
@@ -124,11 +122,22 @@ public class OpaqueByteRangeTest extends StorageAdapterByteRangeTest {
 
                 @Override
                 public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-                    Files.delete(dir);
+                    if (!ofsAdapter.contentPath.equals(dir)) {
+                        Files.delete(dir);
+                    }
                     return FileVisitResult.CONTINUE;
                 }
             });
         }
-        log.info("cleanupBefore: delete all content from " + ofsAdapter.contentPath + " DONE");
+        if (Files.exists(ofsAdapter.txnPath)) {
+            Files.walkFileTree(ofsAdapter.txnPath, new SimpleFileVisitor<Path>() {
+                @Override
+                public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                    Files.delete(file);
+                    return FileVisitResult.CONTINUE;
+                }
+            });
+        }
+        log.info("cleanupBefore: " + ofsAdapter.contentPath.getParent() + " DONE");
     }
 }
